Real-or-Fake-Image-Classifier/Frontend/requirements.txt
streamlit
pillow
requests
python-multipart

